initial

+++ app/src/main/java/com/mobile/agri10x/models/GetOrderListDatumCheckout.java
@@ -10,7 +10,7 @@ public class GetOrderListDatumCheckout {
     private String id;
     @SerializedName("Order_Amount")
     @Expose
-    private Integer orderAmount;
+    private double orderAmount;
     @SerializedName("Order_Notes")
     @Expose
     private String orderNotes;
@@ -62,11 +62,11 @@ public class GetOrderListDatumCheckout {
         this.id = id;
     }
 
-    public Integer getOrderAmount() {
+    public double getOrderAmount() {
         return orderAmount;
     }
 
-    public void setOrderAmount(Integer orderAmount) {
+    public void setOrderAmount(double orderAmount) {
         this.orderAmount = orderAmount;
     }
